Raise a BadRequest if the image id is invalid
Invalid in this case means it's neither an integer nor "front" or
"back". This was previously supposed to raise a NotFound but as there
can't be an image with an invalid id, a BadRequest makes more sense.
(The MusicBrainz server does also use 400 Bad Request if invalid values,
like invalid includes, are supplied.)

This also fixes CAA-58.