Remove Bootstrap
I don't see any good reason to use Bootstrap (pulling in hundreds of
kilobytes of JS/CSS) just to display a single static page that links to
documentation.

I've kept the display of the page nearly identical. The only real
feature of BS we used was the navbar collapse. It no longer collapses
(and doesn't need to, as there are only three items), but does re-orient
itself below the main logo.

This reduces the project complexity significantly (dropping dependencies
on Bootstrap and less, both of which were severly out of date, plus node
and npm), and reduces the unminified page size from ~243 KB to ~71 KB.